refactor: update ktlint to latest version
