+++ .github/workflows/lint.yml
@@ -46,4 +46,4 @@ jobs:
     
     # The task will simply fail if it does not pass. It would be better if we make a comment on the PR telling the PR author to install git hooks but that scenario should not happen often (at least I don't think so) so for now, let's lean on assuming it will not happen and if it does happen a lot, we will improve this. 
     - name: Run ktlint
-      run: make lint 
+      run: make lint-no-format
