+++ common-test/src/main/java/io/customer/commontest/BaseTest.kt
@@ -1,10 +1,10 @@
-package io.customer.common_test
+package io.customer.commontest
 
 import android.app.Application
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.platform.app.InstrumentationRegistry
-import io.customer.common_test.util.DispatchersProviderStub
+import io.customer.commontest.util.DispatchersProviderStub
 import io.customer.sdk.CustomerIOConfig
 import io.customer.sdk.data.model.Region
 import io.customer.sdk.data.store.DeviceStore
