+++ sdk/src/main/java/io/customer/sdk/data/model/EventType.kt
@@ -4,5 +4,5 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = false)
 enum class EventType {
-    event, screen
+    event, screen; // ktlint-disable enum-entry-name-case (enum cases used for JSON values)
 }
