+++ sdk/src/main/java/io/customer/sdk/queue/taskdata/RegisterPushNotificationQueueTaskData.kt
@@ -1,8 +1,8 @@
 package io.customer.sdk.queue.taskdata
 
-import java.util.*
 import com.squareup.moshi.JsonClass
 import io.customer.sdk.data.request.Device
+import java.util.*
 
 @JsonClass(generateAdapter = true)
 internal data class RegisterPushNotificationQueueTaskData(
