+++ sdk/src/sharedTest/java/io/customer/sdk/data/store/DeviceStoreTest.kt
@@ -1,7 +1,7 @@
 package io.customer.sdk.data.store
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.customer.common_test.BaseTest
+import io.customer.commontest.BaseTest
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Test
 import org.junit.runner.RunWith
