+++ sdk/src/sharedTest/java/io/customer/sdk/repository/TrackRepositoryTest.kt
@@ -1,7 +1,7 @@
 package io.customer.sdk.repository
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.customer.common_test.BaseTest
+import io.customer.commontest.BaseTest
 import io.customer.sdk.data.model.EventType
 import io.customer.sdk.data.request.MetricEvent
 import io.customer.sdk.queue.Queue
