+++ sdk/src/sharedTest/java/io/customer/sdk/util/JsonAdapterTest.kt
@@ -1,8 +1,8 @@
 package io.customer.sdk.util
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.customer.common_test.BaseTest
 import com.squareup.moshi.JsonClass
+import io.customer.commontest.BaseTest
 import org.amshove.kluent.AnyException
 import org.amshove.kluent.invoking
 import org.amshove.kluent.shouldBeEqualTo
